Tests compile and basic functionality restored
The parser now annotates alls declarations in the AST with a declaration
site. This site is unique for the translation unit and stable when the
unit is reparsed. It can also be used to locate a declaration in the AST
performantly, because it is encoded as a path through the AST.

The site data is included in the Symbol datastructure, so that it now
truly identifies a symbol uniquely (e.g., it can distinguish between
different declarations of the same type or function).

The dependency analysis is now based on libclang's name resolution
capabilities. Dependencies are finegrained: they resolve to a specific
declaration, not necessarily to the canonical one.