TC-3128 Fix unclear bracketing contract in AstWriters
AstWriters got a statement pretty-printer, but statements in
general are not guaranteed to be bracketed. So when we use this
print a function body (which is a mere Stmt, not Stmt.Block), it
is only guaranteed to be syntactically correct if we add {}.

We used to add those brackets in AstWriters, but that clashed with
the case where the string was coming from source extraction instead
of the pretty-printer.

To fix this, AstWriters now expects a function body pretty printer,
which is promising to include the brackets. Hence, ASTWriters does
not need to add brackets itself.