TC-3136 platform-agnostic model of stdlib
We have different sourcecode generation scenarios that
require different assumptions about standard library definitions.

In some scenarios, we know the environment and want to use type definitions
(e.g. size_t) or malloc for that particular environment, without
depending on some header.

In other scenarios, we do not know the host, and want to use the
agnostic/opaque typedefs/declarations suitable for the environment.

We now support both by abstracting over a suitable model IStdlib of the
C standard library.