Add type dependency analysis, and type resolution
Given a C-type T, we can now compute a closure of sufficient type
dependencies, taking into account that not every type mention requires
a complete definition.

The result of this analysis can be resolved in a type environment,
producing a set of resolutions.

This set of type resolutions can be converted to a preamble that should
generate the necessary type context for T to be valid.