TC-3248 Incrementalization of the LinkAnalysis (now LinkResolution)
We had two link resolution analyses:
- LinkAnalysis
- UnitClosureAnalysis

The former was a whole-program analysis, analyzing a set of units
in one go. The latter was a analysis that incrementally tried to
find unambiguous definitions and build a set of units that maximized
definedness of symbols.

These two should coincide in the sense that any resulting set of the
UnitClosureAnalysis that would be passed to LinkAnalysis should
produce the same LinkGraph.

This was not the case, for a number of reasons (see TC-3248)
The UnitClosureAnalysis was also very specific for benchmarks, because
it took into account minimization.

To solve this, we unified the two analyses and factored out minimization.
We now have a single LinkResolution analysis that models a link target
and can incrementally add more parsed units to that target.
It maintains its state and can output a LinkGraph at any point to
reflect the current status.

The remaining logic that used to be in the closure analysis (such as
consulting the index of missing definitions etc) has been moved
to a new LinkAssistant in truecode that iteratively links more units
using LinkResolution.